CNDB-15001: Rearrange vector key-restricted tests to prevent timeouts
Also remove duplicated rangeSearchTest
