+++ test/unit/org/apache/cassandra/index/sai/cql/VectorKeyRestrictedOnPartitionTest.java
@@ -18,7 +18,7 @@
 
 package org.apache.cassandra.index.sai.cql;
 
-import java.util.Arrays;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -28,20 +28,19 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import org.apache.cassandra.config.CassandraRelevantProperties;
-import org.apache.cassandra.cql3.UntypedResultSet;
+import org.apache.cassandra.index.sai.plan.QueryController;
 import org.apache.cassandra.index.sai.utils.Glove;
-import org.assertj.core.data.Percentage;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class VectorPartitionRestrictedTest extends VectorTester.VersionedWithChecksums
+public class VectorKeyRestrictedOnPartitionTest extends VectorKeyRestrictedTester
 {
     private static Glove.WordVector word2vec;
 
     @BeforeClass
     public static void loadModel() throws Throwable
     {
-        word2vec = Glove.parse(VectorPartitionRestrictedTest.class.getClassLoader().getResourceAsStream("glove.3K.50d.txt"));
+        word2vec = Glove.parse(VectorKeyRestrictedOnPartitionTest.class.getClassLoader().getResourceAsStream("glove.3K.50d.txt"));
     }
 
     @After
@@ -169,42 +168,41 @@ public class VectorPartitionRestrictedTest extends VectorTester.VersionedWithChe
         }
     }
 
-    private void searchWithNonExistingKey(float[] queryVector, int key)
+    @Test
+    public void testPartitionKeyRestrictionCombinedWithSearchPredicate() throws Throwable
     {
-        searchWithKey(queryVector, key, 0);
-    }
+        // Need to test the search then order path
+        QueryController.QUERY_OPT_LEVEL = 0;
 
-    private void searchWithNonExistingKey(Vector<Float> queryVector, int key)
-    {
-        searchWithKey(queryVector, key, 0);
-    }
+        // We use a clustered primary key to simplify the mental model for this test.
+        // The bug this test exposed happens when the last row(s) in a segment, based on PK order, are present
+        // in a peer index for an sstable's search index but not its vector index.
+        createTable("CREATE TABLE %s (partition int, i int, v vector<float, 2>, c int, PRIMARY KEY(partition, i))");
+        createIndex("CREATE CUSTOM INDEX ON %s(v) USING 'StorageAttachedIndex' WITH OPTIONS = {'similarity_function': 'euclidean'}");
+        createIndex("CREATE CUSTOM INDEX ON %s(c) USING 'StorageAttachedIndex'");
 
-    private void searchWithKey(float[] queryVector, int key, int expectedSize)
-    {
-        UntypedResultSet result = execute("SELECT * FROM %s WHERE pk = " + key + " ORDER BY val ann of " + Arrays.toString(queryVector) + " LIMIT 1000");
+        var partitionKeys = new ArrayList<Integer>();
+        // Insert many rows
+        for (int i = 1; i < 1000; i++)
+        {
+            execute("INSERT INTO %s (partition, i, v, c) VALUES (?, ?, ?, ?)", i, i, vector(i, i), i);
+            partitionKeys.add(i);
+        }
 
-        // VSTODO maybe we should have different methods for these cases
-        if (expectedSize < 10)
-            assertThat(result).hasSize(expectedSize);
-        else
-            assertThat(result.size()).isCloseTo(expectedSize, Percentage.withPercentage(5));
-        result.stream().forEach(row -> assertThat(row.getInt("pk")).isEqualTo(key));
-    }
+        beforeAndAfterFlush(() -> {
+            // Restricted by partition key and with low as well as high cardinality of results for column c
+            assertRows(execute("SELECT i FROM %s WHERE partition = 1 AND c > 0 ORDER BY v ANN OF [1,1] LIMIT 1"), row(1));
+            assertRows(execute("SELECT i FROM %s WHERE partition = 1 AND c < 10 ORDER BY v ANN OF [1,1] LIMIT 1"), row(1));
 
-    private void searchWithKey(Vector<Float> queryVector, int key, int expectedSize)
-    {
-        searchWithKey(queryVector, key, expectedSize, 1000);
-    }
+            // Do some partition key range queries, the restriction on c is meaningless, but forces the search then
+            // order path
+            var r1 = execute("SELECT partition FROM %s WHERE token(partition) < token(11) AND c > 0 ORDER BY v ANN OF [1,1] LIMIT 1000");
+            var e1 = keysWithUpperBound(partitionKeys, 11,false);
+            assertThat(keys(r1)).containsExactlyInAnyOrderElementsOf(e1);
 
-    private void searchWithKey(Vector<Float> queryVector, int key, int expectedSize, int limit)
-    {
-        UntypedResultSet result = execute("SELECT * FROM %s WHERE pk = ? ORDER BY val ann of ? LIMIT " + limit, key, queryVector);
-
-        // VSTODO maybe we should have different methods for these cases
-        if (expectedSize < 10)
-            assertThat(result).hasSize(expectedSize);
-        else
-            assertThat(result.size()).isCloseTo(expectedSize, Percentage.withPercentage(10));
-        result.stream().forEach(row -> assertThat(row.getInt("pk")).isEqualTo(key));
+            var r2 = execute("SELECT partition FROM %s WHERE token(partition) >= token(11) AND token(partition) <= token(20) AND c <= 1000 ORDER BY v ANN OF [1,1] LIMIT 1000");
+            var e2 = keysInBounds(partitionKeys, 11, true, 20, true);
+            assertThat(keys(r2)).containsExactlyInAnyOrderElementsOf(e2);
+        });
     }
 }
