+++ test/unit/org/apache/cassandra/index/sai/cql/VectorKeyRestrictedOnRangeTest.java
@@ -18,27 +18,17 @@
 
 package org.apache.cassandra.index.sai.cql;
 
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.junit.Test;
 
 import org.apache.cassandra.cql3.UntypedResultSet;
-import org.apache.cassandra.db.marshal.Int32Type;
-import org.apache.cassandra.dht.IPartitioner;
-import org.apache.cassandra.dht.Murmur3Partitioner;
-import org.apache.cassandra.dht.Token;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class VectorRangeSearchTest extends VectorTester.VersionedWithChecksums
+public class VectorKeyRestrictedOnRangeTest extends VectorKeyRestrictedTester
 {
-    private static final IPartitioner partitioner = Murmur3Partitioner.instance;
-
     @Test
     public void rangeSearchTest() throws Throwable
     {
@@ -56,7 +46,7 @@ public class VectorRangeSearchTest extends VectorTester.VersionedWithChecksums
         }
 
         var queryVector = vector(1.5f, 1.5f);
-        CheckedFunction tester = () -> {
+        beforeAndAfterFlush(() -> {
             for (int i = 1; i <= nPartitions; i++)
             {
                 UntypedResultSet result = execute("SELECT partition FROM %s WHERE token(partition) > token(?) ORDER BY val ann of ? LIMIT 1000", i, queryVector);
@@ -86,49 +76,6 @@ public class VectorRangeSearchTest extends VectorTester.VersionedWithChecksums
                     assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysInBounds(vectorsByKey.keySet(), i, false, j, false));
                 }
             }
-        };
-
-        tester.apply();
-
-        flush();
-
-        tester.apply();
-    }
-
-    private Collection<Integer> keys(UntypedResultSet result)
-    {
-        List<Integer> keys = new ArrayList<>(result.size());
-        for (UntypedResultSet.Row row : result)
-            keys.add(row.getInt("partition"));
-        return keys;
-    }
-
-    private Collection<Integer> keysWithLowerBound(Collection<Integer> keys, int leftKey, boolean leftInclusive)
-    {
-        return keysInTokenRange(keys, partitioner.getToken(Int32Type.instance.decompose(leftKey)), leftInclusive,
-                                partitioner.getMaximumToken().getToken(), true);
-    }
-
-    private Collection<Integer> keysWithUpperBound(Collection<Integer> keys, int rightKey, boolean rightInclusive)
-    {
-        return keysInTokenRange(keys, partitioner.getMinimumToken().getToken(), true,
-                                partitioner.getToken(Int32Type.instance.decompose(rightKey)), rightInclusive);
-    }
-
-    private Collection<Integer> keysInBounds(Collection<Integer> keys, int leftKey, boolean leftInclusive, int rightKey, boolean rightInclusive)
-    {
-        return keysInTokenRange(keys, partitioner.getToken(Int32Type.instance.decompose(leftKey)), leftInclusive,
-                                partitioner.getToken(Int32Type.instance.decompose(rightKey)), rightInclusive);
-    }
-
-    private Collection<Integer> keysInTokenRange(Collection<Integer> keys, Token leftToken, boolean leftInclusive, Token rightToken, boolean rightInclusive)
-    {
-        long left = leftToken.getLongValue();
-        long right = rightToken.getLongValue();
-        return keys.stream()
-                   .filter(k -> {
-                       long t = partitioner.getToken(Int32Type.instance.decompose(k)).getLongValue();
-                       return (left < t || left == t && leftInclusive) && (t < right || t == right && rightInclusive);
-                   }).collect(Collectors.toSet());
+        });
     }
 }
