+++ test/unit/org/apache/cassandra/index/sai/cql/VectorTypeTest.java
@@ -19,14 +19,12 @@
 package org.apache.cassandra.index.sai.cql;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import java.util.stream.Collectors;
 
 import org.junit.Assert;
 import org.junit.BeforeClass;
@@ -35,10 +33,6 @@ import org.junit.Test;
 import io.github.jbellis.jvector.graph.GraphSearcher;
 import io.github.jbellis.jvector.vector.VectorSimilarityFunction;
 import org.apache.cassandra.cql3.UntypedResultSet;
-import org.apache.cassandra.db.marshal.Int32Type;
-import org.apache.cassandra.dht.IPartitioner;
-import org.apache.cassandra.dht.Murmur3Partitioner;
-import org.apache.cassandra.dht.Token;
 import org.apache.cassandra.index.sai.StorageAttachedIndex;
 import org.apache.cassandra.index.sai.disk.v1.IndexWriterConfig;
 import org.apache.cassandra.index.sai.disk.v1.SegmentBuilder;
@@ -60,8 +54,6 @@ import static org.junit.Assert.assertTrue;
 
 public class VectorTypeTest extends VectorTester.Versioned
 {
-    private static final IPartitioner partitioner = Murmur3Partitioner.instance;
-
     @BeforeClass
     public static void setUpClass()
     {
@@ -596,99 +588,6 @@ public class VectorTypeTest extends VectorTester.Versioned
     }
 
     @Test
-    public void rangeSearchTest() throws Throwable
-    {
-        createTable("CREATE TABLE %s (partition int, val vector<float, 2>, PRIMARY KEY(partition))");
-        createIndex("CREATE CUSTOM INDEX ON %s(val) USING 'StorageAttachedIndex' WITH OPTIONS = {'similarity_function' : 'euclidean'}");
-
-        var nPartitions = 100;
-        Map<Integer, float[]> vectorsByKey = new HashMap<>();
-
-        for (int i = 1; i <= nPartitions; i++)
-        {
-            float[] vector = {(float) i, (float) i};
-            execute("INSERT INTO %s (partition, val) VALUES (?, ?)", i, vector(vector));
-            vectorsByKey.put(i, vector);
-        }
-
-        var queryVector = vector(1.5f, 1.5f);
-        CheckedFunction tester = () -> {
-            for (int i = 1; i <= nPartitions; i++)
-            {
-                UntypedResultSet result = execute("SELECT partition FROM %s WHERE token(partition) > token(?) ORDER BY val ann of ? LIMIT 1000", i, queryVector);
-                assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysWithLowerBound(vectorsByKey.keySet(), i, false));
-
-                result = execute("SELECT partition FROM %s WHERE token(partition) >= token(?) ORDER BY val ann of ? LIMIT 1000", i, queryVector);
-                assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysWithLowerBound(vectorsByKey.keySet(), i, true));
-
-                result = execute("SELECT partition FROM %s WHERE token(partition) < token(?) ORDER BY val ann of ? LIMIT 1000", i, queryVector);
-                assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysWithUpperBound(vectorsByKey.keySet(), i, false));
-
-                result = execute("SELECT partition FROM %s WHERE token(partition) <= token(?) ORDER BY val ann of ? LIMIT 1000", i, queryVector);
-                assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysWithUpperBound(vectorsByKey.keySet(), i, true));
-
-                for (int j = 1; j <= nPartitions; j++)
-                {
-                    result = execute("SELECT partition FROM %s WHERE token(partition) >= token(?) AND token(partition) <= token(?) ORDER BY val ann of ? LIMIT 1000", i, j, queryVector);
-                    assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysInBounds(vectorsByKey.keySet(), i, true, j, true));
-
-                    result = execute("SELECT partition FROM %s WHERE token(partition) > token(?) AND token(partition) <= token(?) ORDER BY val ann of ? LIMIT 1000", i, j, queryVector);
-                    assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysInBounds(vectorsByKey.keySet(), i, false, j, true));
-
-                    result = execute("SELECT partition FROM %s WHERE token(partition) >= token(?) AND token(partition) < token(?) ORDER BY val ann of ? LIMIT 1000", i, j, queryVector);
-                    assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysInBounds(vectorsByKey.keySet(), i, true, j, false));
-
-                    result = execute("SELECT partition FROM %s WHERE token(partition) > token(?) AND token(partition) < token(?) ORDER BY val ann of ? LIMIT 1000", i, j, queryVector);
-                    assertThat(keys(result)).containsExactlyInAnyOrderElementsOf(keysInBounds(vectorsByKey.keySet(), i, false, j, false));
-                }
-            }
-        };
-
-        tester.apply();
-
-        flush();
-
-        tester.apply();
-    }
-
-    private Collection<Integer> keys(UntypedResultSet result)
-    {
-        List<Integer> keys = new ArrayList<>(result.size());
-        for (UntypedResultSet.Row row : result)
-            keys.add(row.getInt("partition"));
-        return keys;
-    }
-
-    private Collection<Integer> keysWithLowerBound(Collection<Integer> keys, int leftKey, boolean leftInclusive)
-    {
-        return keysInTokenRange(keys, partitioner.getToken(Int32Type.instance.decompose(leftKey)), leftInclusive,
-                                partitioner.getMaximumToken().getToken(), true);
-    }
-
-    private Collection<Integer> keysWithUpperBound(Collection<Integer> keys, int rightKey, boolean rightInclusive)
-    {
-        return keysInTokenRange(keys, partitioner.getMinimumToken().getToken(), true,
-                                partitioner.getToken(Int32Type.instance.decompose(rightKey)), rightInclusive);
-    }
-
-    private Collection<Integer> keysInBounds(Collection<Integer> keys, int leftKey, boolean leftInclusive, int rightKey, boolean rightInclusive)
-    {
-        return keysInTokenRange(keys, partitioner.getToken(Int32Type.instance.decompose(leftKey)), leftInclusive,
-                                partitioner.getToken(Int32Type.instance.decompose(rightKey)), rightInclusive);
-    }
-
-    private Collection<Integer> keysInTokenRange(Collection<Integer> keys, Token leftToken, boolean leftInclusive, Token rightToken, boolean rightInclusive)
-    {
-        long left = leftToken.getLongValue();
-        long right = rightToken.getLongValue();
-        return keys.stream()
-               .filter(k -> {
-                   long t = partitioner.getToken(Int32Type.instance.decompose(k)).getLongValue();
-                   return (left < t || left == t && leftInclusive) && (t < right || t == right && rightInclusive);
-               }).collect(Collectors.toSet());
-    }
-
-    @Test
     public void selectFloatVectorFunctions()
     {
         createTable(KEYSPACE, "CREATE TABLE %s (pk int primary key, value vector<float, 2>)");
@@ -1048,44 +947,6 @@ public class VectorTypeTest extends VectorTester.Versioned
         assertRows(future.get(), row(2));
 
         assertEquals(0, executor.shutdownNow().size());
-    }
-
-    @Test
-    public void testPartitionKeyRestrictionCombinedWithSearchPredicate() throws Throwable
-    {
-        // Need to test the search then order path
-        QueryController.QUERY_OPT_LEVEL = 0;
-
-        // We use a clustered primary key to simplify the mental model for this test.
-        // The bug this test exposed happens when the last row(s) in a segment, based on PK order, are present
-        // in a peer index for an sstable's search index but not its vector index.
-        createTable("CREATE TABLE %s (partition int, i int, v vector<float, 2>, c int, PRIMARY KEY(partition, i))");
-        createIndex("CREATE CUSTOM INDEX ON %s(v) USING 'StorageAttachedIndex' WITH OPTIONS = {'similarity_function': 'euclidean'}");
-        createIndex("CREATE CUSTOM INDEX ON %s(c) USING 'StorageAttachedIndex'");
-
-        var partitionKeys = new ArrayList<Integer>();
-        // Insert many rows
-        for (int i = 1; i < 1000; i++)
-        {
-            execute("INSERT INTO %s (partition, i, v, c) VALUES (?, ?, ?, ?)", i, i, vector(i, i), i);
-            partitionKeys.add(i);
-        }
-
-        beforeAndAfterFlush(() -> {
-            // Restricted by partition key and with low as well as high cardinality of results for column c
-            assertRows(execute("SELECT i FROM %s WHERE partition = 1 AND c > 0 ORDER BY v ANN OF [1,1] LIMIT 1"), row(1));
-            assertRows(execute("SELECT i FROM %s WHERE partition = 1 AND c < 10 ORDER BY v ANN OF [1,1] LIMIT 1"), row(1));
-
-            // Do some partition key range queries, the restriction on c is meaningless, but forces the search then
-            // order path
-            var r1 = execute("SELECT partition FROM %s WHERE token(partition) < token(11) AND c > 0 ORDER BY v ANN OF [1,1] LIMIT 1000");
-            var e1 = keysWithUpperBound(partitionKeys, 11,false);
-            assertThat(keys(r1)).containsExactlyInAnyOrderElementsOf(e1);
-
-            var r2 = execute("SELECT partition FROM %s WHERE token(partition) >= token(11) AND token(partition) <= token(20) AND c <= 1000 ORDER BY v ANN OF [1,1] LIMIT 1000");
-            var e2 = keysInBounds(partitionKeys, 11, true, 20, true);
-            assertThat(keys(r2)).containsExactlyInAnyOrderElementsOf(e2);
-        });
     }
 
     @Test
